% Generated by roxygen2 (4.0.2): do not edit by hand
\name{accrate.age}
\alias{accrate.age}
\title{Accumulation rate calculation and plotting.

Calculate the accumulation rate (slope of a straight curve between depths) above and below the desired point. Requires sufficiently dense density of depths, e.g. steps=1
to calculate accumulation rate at an age. Before doing this, run your core in clam and store the data, so, provide the option storedat=TRUE
This function is internal to Clam.}
\usage{
accrate.age(age, yrcm = TRUE, prob = 0.95)
}
\arguments{
\item{age}{numeric.  The calibrated year.}

\item{yrcm}{boolean. The estimated error for the <sup>14</sup>C age.}

\item{prob}{numeric. The confidence level (0 - 1).}
}
\value{
A plot of the accumulation rate by age.
}
\description{
Accumulation rate calculation and plotting.

Calculate the accumulation rate (slope of a straight curve between depths) above and below the desired point. Requires sufficiently dense density of depths, e.g. steps=1
to calculate accumulation rate at an age. Before doing this, run your core in clam and store the data, so, provide the option storedat=TRUE
This function is internal to Clam.
}

