% Generated by roxygen2 (4.0.2): do not edit by hand
\name{accrate.depth}
\alias{accrate.depth}
\title{Accumulation rate calculation and plotting.}
\usage{
accrate.depth(depth, yrcm = TRUE, prob = 0.95)
}
\arguments{
\item{depth}{numeric.  The radiocarbon year.}

\item{yrcm}{boolean. The estimated error for the <sup>14</sup>C age.}

\item{prob}{numeric. The confidence level (0 - 1).}
}
\value{
A plot of the accumulation rate by depth.
}
\description{
Calculate the slope of a straight curve between depths above and below the desired point (for each iteration). Requires sufficiently dense density of depths, e.g. yrsteps=1
to calculate accumulation rates at a depth. Before running this, run your core in clam and store the data, so, provide the option storedat=TRUE#' This function is internal to Clam.
}

