% Generated by roxygen2 (4.0.1): do not edit by hand
\name{age.pMC}
\alias{age.pMC}
\title{Calculate pMC values from 14C ages.}
\usage{
age.pMC(mn, sdev, ratio = 100, decimals = 3)
}
\arguments{
\item{mn}{numeric.}

\item{sdev}{numeric.}

\item{ratio}{numeric.}

\item{decimals}{numeric.}
}
\value{
Rounded values for ???mn??? to a specified decimal level.
}
\description{
Calculate the slope of a straight curve between depths above and below the desired point (for each iteration). Requires sufficiently dense density of depths, e.g. yrsteps=1
to calculate accumulation rates at a depth. Before running this, run your core in clam and store the data, so, provide the option storedat=TRUE#' This function is internal to Clam.
}

