% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calibrate}
\alias{calibrate}
\title{Find the calibrated distributions of 14C dates}
\usage{
calibrate(cage = 2450, error = 50, reservoir = 0, prob = 0.95, cc = 1,
  cc1 = "IntCal13.14C", cc2 = "Marine13.14C", cc3 = "SHCal13.14C",
  cc4 = "mixed.14C", cc5 = "gluedHemispheres.14C", postbomb = FALSE,
  pb1 = "postbomb_NH1.14C", pb2 = "postbomb_NH2.14C",
  pb3 = "postbomb_NH3.14C", pb4 = "postbomb_SH1-2.14C",
  pb5 = "postbomb_SH3.14C", yrsteps = 1, pbsteps = 0.01, hpdsteps = 1,
  calibt = FALSE, yrmin = c(), yrmax = c(), minC14 = c(),
  maxC14 = c(), times = 5, calheight = 0.3, expand = 0.1,
  threshold = 1e-06, storedat = FALSE, graph = TRUE, xlab = c(),
  ylab = c(), BCAD = FALSE, mar = c(3.5, 3, 2, 1), mgp = c(2, 1, 0),
  bty = "l", title = c(), date.col = "red", cc.col = rgb(0, 0.5, 0,
  0.7), dist.col = rgb(0, 0, 0, 0.3), sd.col = rgb(0, 0, 0, 0.5))
}
\arguments{
\item{cage}{numeric. The uncalibrated age.}

\item{error}{numeric.  The lab error on the radiocarbon date.}

\item{reservoir}{numeric.  Resevoir effect.}

\item{prob}{numeric. Value from 0 - 1, the significance level.}

\item{cc}{numeric.  Which calibration curve should be used [0 - 5].}

\item{cc1="IntCal13.14C"}{}

\item{cc2="Marine13.14C"}{}

\item{cc3="SHCal13.14C"}{}

\item{cc4="mixed.14C"}{}

\item{cc5="gluedHemispheres.14C"}{}

\item{postbomb}{boolean.  Do we use the post-bomb curve?}

\item{pb1="postbomb_NH1.14C"}{}

\item{pb2="postbomb_NH2.14C"}{}

\item{pb3="postbomb_NH3.14C"}{}

\item{pb4="postbomb_SH1-2.14C"}{}

\item{pb5="postbomb_SH3.14C"}{}

\item{yrsteps=1}{}

\item{pbsteps=0.01}{}

\item{hpdsteps=1}{}

\item{calibt=FALSE}{}

\item{yrmin=c()}{}

\item{yrmax=c()}{}

\item{minC14=c()}{}

\item{maxC14=c()}{}

\item{times=5}{}

\item{calheight=0.3}{}

\item{expand=0.1}{}

\item{threshold=1e-6}{}

\item{storedat=FALSE}{}

\item{graph=TRUE}{}

\item{xlab=c()}{}

\item{ylab=c()}{}

\item{BCAD=FALSE}{}

\item{mar=c(3.5,3,2,1)}{}

\item{mgp=c(2,1,0)}{}

\item{bty="l"}{}

\item{title=c()}{}

\item{date.col="red"}{}

\item{cc.col=rgb(0,.5,0,0.7)}{}

\item{dist.col=rgb(0,0,0,0.3)}{}

\item{sd.col=rgb(0,0,0,0.5)}{}
}
\description{
Find the calibrated distributions of 14C dates
}

