% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam-package.R
\docType{package}
\name{clam-package}
\alias{clam-package}
\title{clam}
\description{
Classical (non-Bayesian) age-depth modelling.
}
\author{
\strong{Maintainer}: Maarten Blaauw \email{maarten.blaauw@qub.ac.uk} (\href{https://orcid.org/0000-0002-5680-1515}{ORCID})

Other contributors:
\itemize{
  \item J. Andres Christen (\href{https://orcid.org/0000-0002-5795-4345}{ORCID}) [contributor]
  \item Judith Esquivel Vazquez [contributor]
  \item Simon Goring [contributor]
}

}
