% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clam}
\alias{clam}
\title{Age model construction.}
\usage{
clam(name = "Example", type = 1, smooth = c(), prob = 0.95,
  its = 1000, wghts = 1, cc = 1, cc.name = c("IntCal13.14C",
  "Marine13.14C", "SHCal13.14C", "mixed.14C", "gluedHemispheres.14C"),
  postbomb = FALSE, pb.name = c("postbomb_NH1.14C", "postbomb_NH2.14C",
  "postbomb_NH3.14C", "postbomb_SH1-2.14C", "postbomb_SH3.14C"),
  outliers = c(), ignore = c(), youngest = c(), extradates = c(),
  slump = c(), est = 1, calibt = FALSE, mixed.effect = FALSE,
  dmin = c(), dmax = c(), every = 1, yrmin = c(), yrmax = c(),
  yrsteps = 1, pbsteps = 0.01, hpdsteps = 1, BCAD = FALSE,
  decimals = 0, accrate = 0, ageofdepth = c(), depth = "cm",
  depthseq = c(), depths.file = FALSE, thickness = 1, hiatus = c(),
  remove.reverse = 0.5, times = 5, sep = ",", ext = ".csv",
  runname = c(), storedat = FALSE, threshold = 1e-06, proxies = FALSE,
  revaxes = FALSE, revd = TRUE, revyr = TRUE, calhght = 0.3,
  maxhght = 0.01, mirror = TRUE, plotrange = TRUE, bty = "l",
  mar = c(3.5, 3, 2, 1), mgp = c(2, 1, 0), plotpdf = TRUE,
  plotpng = TRUE, greyscale = c(), yrlab = c(), dlab = c(),
  calcol = rgb(0, 0.5, 0.5, 0.5), C14col = rgb(0, 0, 1, 0.5),
  outcol = "red", outlsize = 1, bestcol = "black", rangecol = rgb(0, 0,
  0, 0.3), slumpcol = grey(0.75), plotname = TRUE, ash = FALSE)
}
\arguments{
\item{name}{character.  The site name, for use with file-based reconstructions.}

\item{type}{numeric.  Model type to generate the chronology.  Default is 1 (linear), also available are 2, polynomial; 3, cubic spline; 4, smooth spline; 5, locally weighted spline (loess).}

\item{smooth}{numeric.  Associated with the \code{type} variable.  See decription for more details.}

\item{prob}{numeric. Probability level for the reported confidence intervals [0 - 1].}

\item{its}{numeric. Number of iterations used to build the chronology.}

\item{wghts}{numeric. Age weighting, either one of: 1, probability weighting; 2, laboratory error.}

\item{cc}{numeric. Define the calibration curve to be used.  One of 1 - 5.}

\item{cc.name}{character. A vector of calibration curve names, 1, cc2, cc3, cc4, cc5 character.  Calibration curve data.}

\item{postbomb}{boolean.  Are some of the radiocarbon dates post-bomb?  Defaults to FALSE.}

\item{pb.name}{character.  Post-bomb calibration curve names, defaults to \code{Post + cc.name}.}

\item{outliers}{A single value or vector index of dates considered to be outliers (counted from the top).}

\item{ignore}{A single value or vector index of dates that should be ignored entirely.}

\item{youngest}{A single value indicating the minimum possible date for the core.}

\item{extradates}{The numeric depth(s) of any extra dating information, associated with a \code{MyCore_XX.txt} file where XX is equal to \code{extradates}}

\item{slump}{Depth intervals to be excized from the core, listed in order from top to bottom.  Multiple slumps listed in sequence: \code{c(top1, bottom1, top2, bottom2)}}

\item{est}{The method used to convert the estimate to a single point value (from 1 - 7).}

\item{calibt}{The parameters for the t-distribution used for calibrating 14C dates, defaults to Gaussian if not provided.}

\item{mixed.effect}{Model error using Heegaard et al (2005) mixed effects method. Defaults to FALSE.  If TRUE the model uses a default \code{its} of 1000.}

\item{dmin}{Minimum depth of the model (either truncate or force extrapolation).}

\item{dmax}{Maximum depth of the model, as above.}

\item{every}{Depth resolution in cm.}

\item{yrmin}{Set the age-axis minimum limit for plotting (\code{xmin}).}

\item{yrmax}{Set the age-axis maximum limit for plotting (\code{xmax}).}

\item{yrsteps}{Temporal resolution for calibration curves in years.  Default is 1.}

\item{pbsteps}{Temporal resolution for post-bomb calibration curves in years. Default is 0.01.}

\item{hpdsteps}{Temporal resolution for the highest posterior ranges in Calendar years.  Default is 1.}

\item{BCAD}{Are the years expressed in BC/AD (\code{TRUE}) or cal years BP (\code{FALSE}).  Default is \code{FALSE}.}

\item{decimals}{Number of decimals for rounding.  Default is 0, may be positive or negative.}

\item{accrate}{Report deposition rate (yrs/cm: \code{0}) or accumulation rate (cm/yr: \code{1})}

\item{ageofdepth}{Numeric or vector of depths for which ages are to be returned in message output.}

\item{depth}{Depth units. Default is \code{'cm'}.}

\item{depthseq}{Depth intervals for estimation.  Defaults to \code{seq(dmin, dmax, by = every)}.  If \code{depths.file = TRUE} will use depths file.}

\item{depths.file}{If the core has an associated depths file \code{TRUE}.  Default is \code{FALSE}.}

\item{thickness}{Thickness of the dated samples.}

\item{hiatus}{The depths of any hiatuses.  A section must have more than two dated samples for extrapolation.}

\item{remove.reverse}{Proportional threshold for model reversals (0-1) before a warning is thrown.}

\item{times}{The breadth of the sampling interval in standard  deviations.  Default is 5 (99.999996 pct. of CI).}

\item{sep}{Separator for the clam age file.  Default is \code{","}.}

\item{ext}{File extension for clam age file.  Default is \code{".csv"}.}

\item{runname}{Used to generate unique names for individual runs.  If empty, the default is \code{"calibrate"}.}

\item{storedat}{Should data be retained in the Global Environment?  Default is \code{FALSE}.}

\item{threshold}{Threshold probability for sampling exclusion fron calibration curves (default \code{1e-6})}

\item{proxies}{If proxies are included (if a \code{name_proxies.csv} file exists), add them to the plot. Default \code{FALSE}.}

\item{revaxes}{Reverse axes, plotting age on the x-axis and depth on the y-axis.  Default \code{FALSE}.}

\item{revd}{Reverse the depth axis, plotting with 0 at the origin.  Default \code{FALSE}.}

\item{revyr}{Reverse the age axis, with 0 at the origin.  Default \code{FALSE}.}

\item{calhght}{Unit height for the calibrated 14C distributions. Default \code{1}.}

\item{maxhght}{Maximum height for the calibrated age distributions. Default \code{0.01}.}

\item{mirror}{Plot calibrated ages as violin plots or as one sided distributions. Default \code{TRUE}.}

\item{plotrange}{Should the confidence ranges of the model be plotted?  Default \code{TRUE}.}

\item{bty}{Should the plot area be bounded? As in \code{par}.}

\item{mar}{Plot margins.  As in \code{par}.}

\item{mgp}{Axis text margins.  As in \code{par}.}

\item{plotpdf}{Output the plot to PDF.}

\item{plotpng}{Output the plot to PNG.}

\item{greyscale}{Cancel the confidence interval plotting, numeric values}

\item{yrlab}{Axis label for the age axis.  Default is \code{"cal BP"} or \code{"AD/BC"}.}

\item{dlab}{Axis label for the depth axis.  Default is \code{"depth (cm)"}.}

\item{calcol}{\code{rgb} fill color for the calibrated dates (polygon fill).}

\item{C14col}{\code{rgb} color for calibrated ranges of dates (polygon outline).}

\item{outcol}{Color for outlier dates.}

\item{outlsize}{Size of outlier dates.}

\item{bestcol}{Plotting color for the "best" model.}

\item{rangecol}{Plotting color for confidence intervals, as \code{rgb}.}

\item{slumpcol}{Plotting color of any slumps, as \code{rgb}.}

\item{plotname}{Should the core name be printed in the plot?}

\item{ash}{Should all distributions be plotted at the same height?}
}
\value{
A \code{data.frame} with depth, point and 95%CI values.  Also returns a text output and saved file to a 'Cores' folder in the working directory.
#'
}
\description{
clam, R code for classical (non-Bayesian) age-depth modelling, this version 2.2
see accompanying manual.html and Blaauw 2010 (Quaternary Geochronology 5: 512-518)
This is the main age-depth modelling function.
The default values can be changed permanently within this file or temporarily when calling clam()
}
\section{Note}{

Currently the method requires you to have a \code{Cores} folder in your working directory.  This will soon be changes.
}
\examples{
\dontrun{
# You need the neotoma package for this example.  If you do not have it installed, uncomment the following lines:
# require(devtools)
# install_github('neotoma', 'rOpenSci')

require(neotoma)
marion.site <- get_site(sitename='Marion Lake\%')
marion.data <- get_dataset(siteid=marion.site$siteid, datasettype = 'pollen')

marion.download <- get_download(marion.data)

#  You need to have a 'Cores' directory in the current working directory.

if(!'Cores' \%in\% list.files(include.dirs=TRUE)){
  dir.create('Cores')
}

write_agefile(marion.download[[1]],chronology = 1,path = '.', corename = 'Marion', cal.prog = 'Clam')

#  Build a model using a smooth spline:
marion.ages <- clam('Marion', type = 4,
                    depthseq=marion.download[[1]]$sample.meta$depth)

# plot the difference between the Neotoma model and the new clam model:
plot(marion.download[[1]]$chronologies$'NAPD 1'[,2], marion.ages[,4],
     xlab = 'NAPD Ages (uncalibrated)', ylab = 'Clam Ages')
abline(0,1)
}
}
\author{
Maarten Blaauw, Simon Goring
}
\references{
Blaauw, M., 2010. Methods and code for 'classical' age-modelling of radiocarbon sequences. Quaternary Geochronology 5, 512-518
}
\keyword{IO}
\keyword{connection}

