% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glue.curves}
\alias{glue.curves}
\title{Bind calibration curves.}
\usage{
glue.curves(nh = "IntCal09.14C", sh = "SHCal04.14C", offset = c(56, 24),
  name = "gluedHemispheres.14C")
}
\arguments{
\item{nh="IntCal09.14C"}{}

\item{sh="SHCal04.14C"}{}

\item{offset=c(56,}{24)}

\item{name="gluedHemispheres.14C"}{}
}
\value{
Outputs a table with a bound calibration curve.
}
\description{
This function 'glues' two curves together, e.g., for calibrating southern hemisphere dates older than SHCal04
}

