% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mix.curves}
\alias{mix.curves}
\title{Create a new calibration curve if dates are of mixed source.}
\usage{
mix.curves(ratio = 0.5, cc1 = "IntCal13.14C", cc2 = "Marine13.14C",
  name = "mixed.14C", offset = c(0, 0))
}
\arguments{
\item{ratio=.5}{}

\item{cc1="IntCal13.14C"}{}

\item{cc2="Marine13.14C"}{}

\item{name="mixed.14C"}{}

\item{offset=c(0,0)}{}
}
\value{
Output to a new table at location defined in \code{name}.
}
\description{
If two curves need to be 'mixed' to calibrate, then re-generate a calibration curve. e.g. for dates of mixed terrestrial and marine carbon sources
}

