% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.proxies}
\alias{plot.proxies}
\title{Plot age model with proxy data.}
\usage{
\method{plot}{proxies}(prox, errors = TRUE, proxcol = grey(0.5),
  revyr = TRUE)
}
\arguments{
\item{prox}{}

\item{errors=TRUE}{}

\item{proxcol=grey(0.5)}{}

\item{revyr=TRUE}{}
}
\value{
A plot of the age model function with proxies.
}
\description{
Only works after doing a clam run with \code{proxies=TRUE}.
}

