% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pMC.age}
\alias{pMC.age}
\title{Calculate 14C ages from pmC values}
\usage{
pMC.age(mn, sdev, ratio = 100, decimals = 0)
}
\arguments{
\item{mn}{}

\item{sdev}{}

\item{ratio=100}{}

\item{decimals=0}{}
}
\value{
Rounded estimates of the 14C dates.
}
\description{
Calculate 14C ages from pmC values
}

