% Generated by roxygen2 (4.0.1): do not edit by hand
\name{student.t}
\alias{student.t}
\title{Use student-t approach to model 14C calibration of a date and return a plot of the calibrated age.}
\usage{
student.t(y = 2450, error = 50, t.a = 3, t.b = 4, cc = 1,
  postbomb = c(), cc1 = "IntCal13", cc2 = "Marine13", cc3 = "SHCal13",
  cc4 = "ConstCal", Cutoff = 1e-05)
}
\arguments{
\item{y}{numeric.  The radiocarbon year.}

\item{error}{numeric. The estimated error for the 14C age.}

\item{t.a}{numeric. The a parameter for the student's t distribution.}

\item{t.b}{numeric. The b parameter for the student's t distribution.}

\item{cc}{numeric. Should this value be calibrated, or is it already calibrated?}

\item{postbomb}{numeric.  A value of 1 - 5 related to the postbomb curve that should be used.  See details.}

\item{cc1}{character. One of IntCal13, Marine13, SHCal13, or ConstCal.}

\item{cc2}{character. As above.}

\item{cc3}{character. As above.}

\item{cc4}{character. As above.}

\item{Cutoff}{numeric.}
}
\value{
A plot of the calibrated age.
}
\description{
Using the caibration curve defined by the user, return the radiocarbon calibration curve (in cal yr BP) for a radiocarbon date.
 Using student's t to model the curve provides greater support to low probability dates at the margins
 of the calibration curve.
}

